+++ src/main/resources/io/viash/helpers/bashutils/ViashDockerAutodetectMount.sh
@@ -1,7 +1,8 @@
 # ViashDockerAutodetectMount: auto configuring docker mounts from parameters
-# $1                  : The parameter value
-# returns             : New parameter
-# $VIASH_DIRECTORY_MOUNTS : Added another parameter to be passed to docker
+# $1                             : The parameter value
+# returns                        : New parameter
+# $VIASH_DIRECTORY_MOUNTS        : Added another parameter to be passed to docker
+# $VIASH_DOCKER_AUTOMOUNT_PREFIX : The prefix to be used for the automounts
 # examples:
 #   ViashDockerAutodetectMount /path/to/bar      # returns '/viash_automount/path/to/bar'
 #   ViashDockerAutodetectMountArg /path/to/bar   # returns '--volume="/path/to:/viash_automount/path/to"'
@@ -14,7 +15,7 @@ function ViashDockerAutodetectMount {
     mount_source=`dirname "$abs_path"`
     base_name=`basename "$abs_path"`
   fi
-  mount_target="/viash_automount$mount_source"
+  mount_target="$VIASH_DOCKER_AUTOMOUNT_PREFIX$mount_source"
   if [ -z "$base_name" ]; then
     echo "$mount_target"
   else
@@ -30,11 +31,11 @@ function ViashDockerAutodetectMountArg {
     mount_source=`dirname "$abs_path"`
     base_name=`basename "$abs_path"`
   fi
-  mount_target="/viash_automount$mount_source"
+  mount_target="$VIASH_DOCKER_AUTOMOUNT_PREFIX$mount_source"
   ViashDebug "ViashDockerAutodetectMountArg $1 -> $mount_source -> $mount_target"
   echo "--volume=\"$mount_source:$mount_target\""
 }
 function ViashDockerStripAutomount {
   abs_path=$(ViashAbsolutePath "$1")
-  echo "${abs_path#/viash_automount}"
+  echo "${abs_path#$VIASH_DOCKER_AUTOMOUNT_PREFIX}"
 }