+++ src/main/scala/io/viash/runners/ExecutableRunner.scala
@@ -101,6 +101,11 @@ final case class ExecutableRunner(
   @default("Empty")
   docker_run_args: OneOrMore[String] = Nil,
 
+  @description("The default prefix to use for automounting directories in Docker. This is used to mount directories from the host into the Docker container.")
+  @example("docker_automount_prefix: /viash_automount", "yaml")
+  @default("/viash_automount")
+  docker_automount_prefix: String = "/viash_automount",
+
   `type`: String = "executable"
 ) extends Runner {
 
@@ -290,7 +295,12 @@ final case class ExecutableRunner(
     }
     
     val preParse =
-      s"""${Bash.ViashDockerFuns}
+      s"""# configure default docker automount prefix if it is unset
+        |if [ -z "$${VIASH_DOCKER_AUTOMOUNT_PREFIX+x}" ]; then
+        |  VIASH_DOCKER_AUTOMOUNT_PREFIX="${docker_automount_prefix}"
+        |fi
+        |
+        |${Bash.ViashDockerFuns}
         |
         |# ViashDockerFile: print the dockerfile to stdout
         |# $$1    : engine identifier